refactor event-edit.js, move all additional template function for different files.
