+++ src/view/event-destination-template.js
@@ -1,0 +1,8 @@
+export const createEventDestinationTemplate = (destination) => {
+  return `<input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="${destination}" list="destination-list-1">
+          <datalist id="destination-list-1">
+            <option value="Amsterdam"></option>
+            <option value="Geneva"></option>
+            <option value="Chamonix"></option>
+          </datalist>`;
+};
