+++ src/view/event-edit.js
@@ -1,121 +1,10 @@
-import {createOffers} from "./event-offers-available.js";
 import SmartView from "./smart.js";
+import {createEventTypeIconTemplate} from "./event-icon-template.js";
+import {createEventTypeTemplate} from "./event-type-template.js";
+import {createEventDestinationTemplate} from "./event-destination-template.js";
+import {createEventEditOffersTemplate} from "./event-offers-template.js";
+import {createEventEditDescriptionTemplate} from "./event-description-template.js";
 import {getEventOffers, getNewInformation, ALL_OFFERS} from "../mock/event.js";
-
-const createEventTypeIconTemplate = (type) => {
-  return `<label class="event__type  event__type-btn" for="event-type-toggle-1">
-            <span class="visually-hidden">Choose event type</span>
-            <img class="event__type-icon" width="17" height="17" src="img/icons/${type.toLowerCase()}.png" alt="Event type icon">
-          </label>
-          <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
-
-          <div class="event__type-list">
-            <fieldset class="event__type-group">
-              <legend class="visually-hidden">Event type</legend>
-
-              <div class="event__type-item">
-                <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
-                <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-1">Taxi</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-bus-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="bus">
-                <label class="event__type-label  event__type-label--bus" for="event-type-bus-1">Bus</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-train-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="train">
-                <label class="event__type-label  event__type-label--train" for="event-type-train-1">Train</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-ship-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="ship">
-                <label class="event__type-label  event__type-label--ship" for="event-type-ship-1">Ship</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-transport-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="transport">
-                <label class="event__type-label  event__type-label--transport" for="event-type-transport-1">Transport</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-drive-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="drive">
-                <label class="event__type-label  event__type-label--drive" for="event-type-drive-1">Drive</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked>
-                <label class="event__type-label  event__type-label--flight" for="event-type-flight-1">Flight</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="check-in">
-                <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-1">Check-in</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-sightseeing-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="sightseeing">
-                <label class="event__type-label  event__type-label--sightseeing" for="event-type-sightseeing-1">Sightseeing</label>
-              </div>
-
-              <div class="event__type-item">
-                <input id="event-type-restaurant-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="restaurant">
-                <label class="event__type-label  event__type-label--restaurant" for="event-type-restaurant-1">Restaurant</label>
-              </div>
-            </fieldset>
-          </div>`;
-};
-
-const createEventTypeTemplate = (type) => {
-  return `<label class="event__label  event__type-output" for="event-destination-1">
-            ${type}
-          </label>`;
-};
-
-const createEventDestinationTemplate = (destination) => {
-  return `<input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="${destination}" list="destination-list-1">
-          <datalist id="destination-list-1">
-            <option value="Amsterdam"></option>
-            <option value="Geneva"></option>
-            <option value="Chamonix"></option>
-          </datalist>`;
-};
-
-const createEventEditOffersTemplate = (offers) => {
-  const offersTemplate = createOffers(offers);
-
-  return `<section class="event__section  event__section--offers">
-            <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-            ${offersTemplate}
-          </section>`;
-};
-
-const createEventEditPhotosTemplate = (photos) => {
-  let photosList = ``;
-
-  photos.forEach((photo) => {
-    photosList += `<img class="event__photo" src="${photo}" alt="Event photo">`;
-  });
-
-  return `<div class="event__photos-container">
-            <div class="event__photos-tape">
-            ${photosList}
-            </div>
-          </div>`;
-};
-
-const createEventEditDescriptionTemplate = (information) => {
-  const description = information.description;
-  const photos = information.photos;
-
-  const photosTemplate = createEventEditPhotosTemplate(photos);
-
-  return `<section class="event__section  event__section--destination">
-            <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-            <p class="event__destination-description">${description}</p>
-            ${photosTemplate}
-          </section>`;
-};
 
 const createEventEditTemplate = (data) => {
   const {type, destination, times, price, offers, information} = data;
@@ -129,44 +18,44 @@ const createEventEditTemplate = (data) => {
   const descriptionTemplate = createEventEditDescriptionTemplate(information);
 
   return `<li class="trip-events__item">
-              <form class="event event--edit" action="#" method="post">
-                <header class="event__header">
-                  <div class="event__type-wrapper">
-                    ${typeIconTemplate}
-                  </div>
-                  <div class="event__field-group  event__field-group--destination">
-                    ${typeTemplate}
-                    ${destinationTemplate}
-                  </div>
+            <form class="event event--edit" action="#" method="post">
+              <header class="event__header">
+                <div class="event__type-wrapper">
+                  ${typeIconTemplate}
+                </div>
+                <div class="event__field-group  event__field-group--destination">
+                  ${typeTemplate}
+                  ${destinationTemplate}
+                </div>
 
-                  <div class="event__field-group  event__field-group--time">
-                    <label class="visually-hidden" for="event-start-time-1">From</label>
-                    <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${startTime}">
-                    &mdash;
-                    <label class="visually-hidden" for="event-end-time-1">To</label>
-                    <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${endTime}">
-                  </div>
+                <div class="event__field-group  event__field-group--time">
+                  <label class="visually-hidden" for="event-start-time-1">From</label>
+                  <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${startTime}">
+                  &mdash;
+                  <label class="visually-hidden" for="event-end-time-1">To</label>
+                  <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${endTime}">
+                </div>
 
-                  <div class="event__field-group  event__field-group--price">
-                    <label class="event__label" for="event-price-1">
-                      <span class="visually-hidden">Price</span>
-                      &euro;
-                    </label>
-                    <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="${price}">
-                  </div>
+                <div class="event__field-group  event__field-group--price">
+                  <label class="event__label" for="event-price-1">
+                    <span class="visually-hidden">Price</span>
+                    &euro;
+                  </label>
+                  <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="${price}">
+                </div>
 
-                  <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-                  <button class="event__reset-btn" type="reset">Delete</button>
-                  <button class="event__rollup-btn" type="button">
-                    <span class="visually-hidden">Open event</span>
-                  </button>
-                </header>
-                <section class="event__details">
-                  ${offerTemplate}
-                  ${descriptionTemplate}
-                </section>
-              </form>
-            </li>`;
+                <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
+                <button class="event__reset-btn" type="reset">Delete</button>
+                <button class="event__rollup-btn" type="button">
+                  <span class="visually-hidden">Open event</span>
+                </button>
+              </header>
+              <section class="event__details">
+                ${offerTemplate}
+                ${descriptionTemplate}
+              </section>
+            </form>
+          </li>`;
 };
 
 export default class EventEdit extends SmartView {
