+++ src/view/event-type-template.js
@@ -1,0 +1,5 @@
+export const createEventTypeTemplate = (type) => {
+  return `<label class="event__label  event__type-output" for="event-destination-1">
+            ${type}
+          </label>`;
+};
