+++ src/view/trip-cost.js
@@ -1,14 +1,18 @@
 import AbstractView from "./abstract-view.js";
 
 const createTripCostTemplate = (events) => {
-  let price = 0;
+  let totalPrice = 0;
 
   for (let i = 0; i < events.length; i++) {
-    price += events[i].price;
+    let offerPrice = 0;
+    events[i].offers.forEach((offer) => {
+      offerPrice += offer.price;
+    });
+    totalPrice += events[i].price + offerPrice;
   }
 
   return `<p class="trip-info__cost">
-              Total: &euro;&nbsp;<span class="trip-info__cost-value">${price}</span>
+              Total: &euro;&nbsp;<span class="trip-info__cost-value">${totalPrice}</span>
           </p>`;
 };
